requirements.txt
# Core ML/AI frameworks
torch>=2.0.0
dgl>=1.0.0

# Parsing
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0
tree-sitter-java>=0.20.0

# Backend & API
fastapi>=0.104.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
uvicorn[standard]>=0.24.0

# Data & Graph processing
networkx>=3.0
neo4j>=5.0.0
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0
